add env config
