+++ .env.dev
@@ -1,0 +1,1 @@
+REACT_APP_HOST="https://contact-tracing-26716.web.app/"
