+++ .env.stg
@@ -1,0 +1,1 @@
+REACT_APP_HOST=""
