+++ firebase.json
@@ -1,10 +1,16 @@
 {
   "hosting": {
     "public": "build",
-    "ignore": [
-      "firebase.json",
-      "**/.*",
-      "**/node_modules/**"
+    "ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
+    "rewrites": [
+      {
+        "source": "/**",
+        "destination": "/index.html"
+      },
+      {
+        "source": "!/@(js|css)/**",
+        "destination": "/index.html"
+      }
     ]
   }
 }
