+++ package.json
@@ -26,14 +26,17 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "sh -ac '. ./.env.${REACT_APP_ENV}; react-scripts build'",
+    "build:dev": "REACT_APP_ENV=dev npm run build",
+    "build:stg": "REACT_APP_ENV=stg npm run build",
+    "build:prd": "REACT_APP_ENV=prd npm run build",
     "test": "react-scripts test",
     "test:cov": "npm run test -- --coverage --watchAll=false",
     "test:e2e": "echo \"need e2e test scripts\"",
     "lint": "eslint --fix 'src/**/*.{ts,tsx}'",
-    "deploy:dev": "npm run build && firebase deploy --token \"$FIREBASE_DEV_TOKEN\"",
-    "deploy:stg": "npm run build && echo \"need deploy scripts\"",
-    "deploy:prd": "npm run build && echo \"need deploy scripts\"",
+    "deploy:dev": "npm run build:dev && firebase deploy --token \"$FIREBASE_DEV_TOKEN\"",
+    "deploy:stg": "npm run build:stg && echo \"need deploy scripts\"",
+    "deploy:prd": "npm run build:prd && echo \"need deploy scripts\"",
     "openapi-generate": "openapi-generator generate -i swagger/openapi.yml -g typescript-axios -o src/apis/openapi",
     "eject": "react-scripts eject"
   },
